Move counter and create showCounter.js module. Call it in index.js. Variable showsTotal created and updateCounter is targeting DOM element nav-counter. Increment is adding +1 for each show. Function counts and updates shows count only after rendering using promises
